+++ setUpServers.cpp
@@ -1,4 +1,4 @@
-#include  "core.hpp"
+#include  "webserv.hpp"
 
 /**
  * @brief this function creat Servers from ServerConfig and setup each server,
